organize repo structure and add .gitignore
